[update]update GameSceneController.cs
commit 26180266328c4c219392d73f4cbe330b566f8192
Author: koki0419 <azunyan40.19@gmail.com>
Date:   Sat Mar 16 03:28:15 2019 +0900

    [update]update GameSceneController.cs

commit 30393dfa1d38d73a485367bf922afcc7d41bcae9
Author: koki0419 <azunyan40.19@gmail.com>
Date:   Sat Mar 16 03:20:23 2019 +0900

    [remove]Remove old player

commit 56ae4338221d381b4583cb22e7e845d46e8ba6f9
Author: koki0419 <azunyan40.19@gmail.com>
Date:   Sat Mar 16 03:15:37 2019 +0900

    [add]add ChargePointManager.cs
    [remove]Remove GameSceneController.cs in chargepointscript
