+++ UnityProjct/Assets/Animetion.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6873b2ebd80d6504793c6c6062322d08
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
