+++ UnityProjct/Assets/ArtAsset/Kuromaru/Materials.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f1dd46449d9db2049a6a1167cd0e1899
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
