+++ UnityProjct/Assets/ArtAsset/Player/2DObject.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3273ca97341f4c848b8a44f838a11913
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
