+++ UnityProjct/Assets/DamageText/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e0ca45773e2e98d4f8b1983f146cc382
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
