+++ UnityProjct/Assets/DamageText/Scripts/Effect.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 00e08dcbcb88c9c4092c158bbbf6d373
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
