+++ UnityProjct/Assets/Scripts/BreakBoxController.cs
@@ -66,7 +66,7 @@ public class BreakBoxController : MonoBehaviour
     {
         foundationHPMax = foundationHP;
         //『PlayerMove』を取得します
-        playerMove = Singleton.Instance.gameSceneController.PlayerMove;
+        playerMove = Singleton.Instance.gameSceneController.playerMove;
         //破壊したときの動き
         onMove = false;
         //オブジェクトを削除するかどうか
