+++ UnityProjct/Assets/Scripts/ChargePointManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 93b22e397f7f60f40bd562c215110756
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
