+++ UnityProjct/Assets/Sound/BGM.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d0c50575a69e5b34eaba5a0f610738d3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
