+++ UnityProjct/ProjectSettings/ProjectVersion.txt
@@ -1,1 +1,1 @@
-m_EditorVersion: 2018.3.7f1
+m_EditorVersion: 2018.3.5f1
